# docs/swagger.yaml
definitions:
  ai.ForwardDirectionRequest:
    properties:
      experimentCondition:
        description: |-
          ExperimentCondition string 实验条件
          步骤对应的实验条件描述
        example: 实验条件
        type: string
      materialGroups:
        description: |-
          MaterialGroups []MaterialGroupData 材料组
          该步骤中涉及的材料组信息
        items:
          $ref: '#/definitions/data.MaterialGroupData'
        type: array
      stepName:
        description: |-
          StepName string 步骤名称
          描述实验步骤的名称
        example: 步骤名称
        type: string
    type: object
  ai.ForwardDirectionResponse:
    properties:
      aiResult:
        allOf:
        - $ref: '#/definitions/ai.ForwardDirectionResult'
        description: |-
          AiResult ForwardDirectionResult 数据
          ai的结果
      historyList:
        description: |-
          historyList []ForwardDirectionResultInfo 数据
          实验记录的结果
        items:
          $ref: '#/definitions/ai.ForwardDirectionResultInfo'
        type: array
    type: object
  ai.ForwardDirectionResult:
    properties:
      resultValue:
        description: |-
          ResultValue string 实验条件
          步骤对应的实验结果
        example: 步骤结果值
        type: string
      stepName:
        description: |-
          StepName string 步骤名称
          描述实验步骤的名称
        example: 步骤名称
        type: string
    type: object
  ai.ForwardDirectionResultInfo:
    properties:
      experimentName:
        description: |-
          ExperimentName string 实验名称
          实验的名称信息
        example: 实验名称
        type: string
      resultValue:
        description: |-
          ResultValue string 实验条件
          步骤对应的实验结果
        example: 步骤结果值
        type: string
      stepName:
        description: |-
          StepName string 步骤名称
          描述实验步骤的名称
        example: 步骤名称
        type: string
    type: object
...
  data.ExperimentAddRequest:
    properties:
      endTime:
        description: |-
          EndTime string 实验结束时间
          选填，格式为 "2006-01-02T15:04:05Z"（RFC3339 格式）
          示例值: "2024-02-05T18:00:00Z"
        example: "2024-02-05T18:00:00Z"
        type: string
      experimentName:
        description: |-
          ExperimentName string 实验名称
          实验的名称，选填；如果填写，名称长度限制为 2-128 字符
          支持中文、英文、数字及特殊字符
          示例值: "实验名称"
        example: 实验名称
        maxLength: 128
        minLength: 2
        type: string
      experimenter:
        description: |-
          Experimenter string 实验者
          实验的负责人，选填；如果填写，名称长度限制为 1-128 字符
          示例值: "张三"
        example: 张三
        maxLength: 128
        minLength: 1
        type: string
      sort:
        description: |-
          Sort int 排序 优先级从大到小
          选填，必须是大于等于 0 的整数
          示例值: 1
        example: 1
        minimum: 0
        type: integer
      startTime:
        description: |-
          StartTime string 实验开始时间
          选填，格式为 "2006-01-02T15:04:05Z"（RFC3339 格式）
          示例值: "2024-02-05T08:30:00Z"
        example: "2024-02-05T08:30:00Z"
        type: string
      steps:
        description: |-
          Steps []ExperimentStepAdd 实验步骤列表
          选填，包含该实验的所有步骤信息，每个步骤包含名称、描述、实验条件、结果值及材料组等
          示例值: [{"stepName": "步骤名称", "stepNameDescription": "实验步骤描述", "experimentCondition": "实验条件", "resultValue": "步骤结果值", "materialGroups": []}]
        items:
          $ref: '#/definitions/data.ExperimentStepAdd'
        type: array
    type: object
...
  data.ExperimentDeleteRequest:
    properties:
      experimentId:
        description: |-
          experimentId 实验编号，必填，UUID格式
          唯一标识要删除的实验，格式必须为UUID4
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  data.ExperimentListRequest:
    properties:
      experimentName:
        description: |-
          ExperimentName string 实验名称，选填，长度限制：2-128字符
          用于按名称模糊查询实验记录
        example: 实验名称
        maxLength: 128
        minLength: 2
        type: string
      experimenter:
        description: |-
          Experimenter string 实验者，选填，长度限制：2-128字符
          进行实验的人员名称
        example: 张三
        maxLength: 128
        minLength: 1
        type: string
      page:
        description: |-
          Page 页码，选填，范围限制：[1,10000]
          用于分页查询管理员列表，最小值为1，最大值为10000
        example: 1
        maximum: 10000
        minimum: 1
        type: integer
      pageSize:
        description: |-
          PageSize 每页大小，选填，范围限制：[1,10000]
          用于限制每页返回的管理员数量，最小值为1，最大值为10000
        example: 10
        maximum: 10000
        minimum: 1
        type: integer
    type: object
  data.ExperimentListResponse:
    properties:
      list:
        description: |-
          List []ExperimentData 数据
          分页返回的实验记录列表
        items:
          $ref: '#/definitions/data.ExperimentData'
        type: array
      total:
        description: |-
          Total int64 总条数
          返回符合条件的实验记录总数
        example: 100
        type: integer
    type: object
...
  data.ExperimentUpdateRequest:
    properties:
      endTime:
        description: |-
          EndTime string 实验结束时间
          选填，格式为 "2006-01-02T15:04:05Z"（RFC3339 格式）
          示例值: "2024-02-05T18:00:00Z"
        example: "2024-02-05T18:00:00Z"
        type: string
      experimentId:
        description: |-
          ExperimentID string 实验ID
          唯一标识实验的 UUID，必填，用于确定要更新的实验记录
          示例值: "123e4567-e89b-12d3-a456-426614174000"
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      experimentName:
        description: |-
          ExperimentName string 实验名称
          实验的名称，选填；如果填写，名称长度限制为 2-128 字符
          支持中文、英文、数字及特殊字符
          示例值: "实验名称"
        example: 实验名称
        maxLength: 128
        minLength: 2
        type: string
      experimenter:
        description: |-
          Experimenter string 实验者
          实验的负责人，选填；如果填写，名称长度限制为 1-128 字符
          示例值: "张三"
        example: 张三
        maxLength: 128
        minLength: 1
        type: string
      startTime:
        description: |-
          StartTime string 实验开始时间
          选填，格式为 "2006-01-02T15:04:05Z"（RFC3339 格式）
          示例值: "2024-02-05T08:30:00Z"
        example: "2024-02-05T08:30:00Z"
        type: string
      steps:
        description: |-
          Steps []ExperimentStepUpdate 实验步骤列表
          选填，包含该实验的所有步骤信息，每个步骤包含名称、描述、实验条件、结果值及材料组等
          示例值: [{"stepName": "步骤名称", "stepNameDescription": "实验步骤描述", "experimentCondition": "实验条件", "resultValue": "步骤结果值", "materialGroups": []}]
        items:
          $ref: '#/definitions/data.ExperimentStepUpdate'
        type: array
    required:
    - experimentId
    type: object
...
  data.MaterialGroupData:
    properties:
      materialGroupName:
        description: |-
          MaterialGroupName string 材料组名称
          材料组的名称信息
        example: 材料组名称
        type: string
      materials:
        description: |-
          Materials []MaterialData 材料列表
          材料组内的具体材料信息
        items:
          $ref: '#/definitions/data.MaterialData'
        type: array
      proportion:
        description: |-
          Proportion float64 材料组占比
          材料组在实验步骤中的占比，百分比形式
        example: 25.5
        type: number
    type: object
  data.MaterialGroupInfo:
    properties:
      materialGroupId:
        description: |-
          MaterialGroupID string 材料组ID
          唯一标识材料组的UUID
        example: 123e4567-e89b-12d3-a456-426614174003
        type: string
      materialGroupName:
        description: |-
          MaterialGroupName string 材料组名称
          材料组的名称信息
        example: 材料组名称
        type: string
      materials:
        description: |-
          Materials []MaterialInfo 材料列表
          材料组内的具体材料信息
        items:
          $ref: '#/definitions/data.MaterialInfo'
        type: array
      proportion:
        description: |-
          Proportion float64 材料组占比
          材料组在实验步骤中的占比，百分比形式
        example: 25.5
        type: number
    type: object
...
  data.RecipeAddRequest:
    properties:
      materialGroups:
        description: |-
          MaterialGroupData []MaterialGroupData 材料组列表
          选填，实验步骤中涉及的材料组信息, percentage的和为100(占比100%)
          示例值: [{"materialGroupName": "材料组名称", "proportion": 25.5, "materials": []}]
        items:
          $ref: '#/definitions/data.MaterialGroupData'
        type: array
      recipeName:
        description: |-
          RecipeName string 配方名称，选填，长度限制：2-255字符
          用于按名称模糊查询实验记录
        example: 配方名称
        maxLength: 255
        minLength: 2
        type: string
      sort:
        description: |-
          Sort int 排序 优先级从大到小
          选填，必须是大于等于 0 的整数
          示例值: 1
        example: 1
        minimum: 0
        type: integer
    type: object
...
  data.RecipeDeleteRequest:
    properties:
      recipeId:
        description: |-
          RecipeId 是配方的唯一标识符
          必填字段，格式为 UUID v4
          示例值: "123e4567-e89b-12d3-a456-426614174000"
        example: 配方id
        type: string
    required:
    - recipeId
    type: object
  data.RecipeEditRequest:
    properties:
      materialGroups:
        description: |-
          MaterialGroupData []MaterialGroupData 材料组列表
          选填，实验步骤中涉及的材料组信息, percentage的和为100(占比100%)
          示例值: [{"materialGroupName": "材料组名称", "proportion": 25.5, "materials": []}]
        items:
          $ref: '#/definitions/data.MaterialGroupInfo'
        type: array
      recipeId:
        description: |-
          RecipeId 是配方的唯一标识符
          必填字段，格式为 UUID v4
          示例值: "123e4567-e89b-12d3-a456-426614174000"
        example: 配方id
        type: string
      recipeName:
        description: |-
          RecipeName string 配方名称，选填，长度限制：2-255字符
          用于按名称模糊查询实验记录
        example: 配方名称
        maxLength: 255
        minLength: 2
        type: string
      sort:
        description: |-
          Sort int 排序 优先级从大到小
          选填，必须是大于等于 0 的整数
          示例值: 1
        example: 1
        minimum: 0
        type: integer
    required:
    - materialGroups
    - recipeId
    - recipeName
    - sort
    type: object
  data.RecipeFormListRequest:
    properties:
      page:
        description: |-
          Page 页码，选填，范围限制：[1,10000]
          用于分页查询管理员列表，最小值为1，最大值为10000
        example: 1
        maximum: 10000
        minimum: 1
        type: integer
      pageSize:
        description: |-
          PageSize 每页大小，选填，范围限制：[1,10000]
          用于限制每页返回的管理员数量，最小值为1，最大值为10000
        example: 10
        maximum: 10000
        minimum: 1
        type: integer
      recipeName:
        description: |-
          RecipeName 是一个可选的字符串字段，用于按配方名称进行模糊查询。
          该字段的长度必须在 2 到 255 个字符之间。如果不提供该字段，则不进行名称过滤。
          示例值："配方名称"，系统会查找名称中包含该字符串的所有配方记录。
        example: 配方名称
        maxLength: 255
        minLength: 2
        type: string
    type: object
  data.RecipeFormListResponse:
    properties:
      list:
        description: |-
          List 是一个 RecipeInfo 类型的切片，用于存储分页返回的配方记录列表。
          例如，列表中可能包含多个不同配方的信息，前端可以根据这些信息展示配方列表。
        items:
          $ref: '#/definitions/data.RecipeInfo'
        type: array
      total:
        description: |-
          Total 表示符合查询条件的配方记录的总条数。
          示例值："100"，表示共有 100 条记录符合查询条件。
        example: 100
        type: integer
    type: object
...
  data.RecipeInfoRequest:
    properties:
      recipeId:
        description: |-
          RecipeId 是配方的唯一标识符
          必填字段，格式为 UUID v4
          示例值: "123e4567-e89b-12d3-a456-426614174000"
        example: 配方id
        type: string
    required:
    - recipeId
    type: object
  data.RecipeInfoResponse:
    properties:
      createdAt:
        description: |-
          CreatedAt 表示配方的创建时间
          时间格式为 ISO 8601 格式的 UTC 时间戳
          示例值: "2024-11-18T10:00:00Z"
        example: "2024-11-18T10:00:00Z"
        type: string
      materialGroups:
        description: |-
          MaterialGroups 包含与该配方关联的材料组信息
          每个材料组都包含详细的材料信息
        items:
          $ref: '#/definitions/data.MaterialGroupInfo'
        type: array
      recipeBasedExperiment:
        description: RecipeBasedExperiment 包含基于此配方创建的实验信息列表
        items:
          $ref: '#/definitions/data.ExperimentInfo'
        type: array
      recipeId:
        description: |-
          RecipeId string 配方id
          UUID
        example: id
        type: string
      recipeName:
        description: |-
          RecipeName 表示配方的名称
          返回配方的名称信息
          示例值: "配方名称"
        example: 配方名称
        type: string
    type: object
  data.RecipeListRequest:
    properties:
      page:
        description: |-
          Page 页码，选填，范围限制：[1,10000]
          用于分页查询管理员列表，最小值为1，最大值为10000
        example: 1
        maximum: 10000
        minimum: 1
        type: integer
      pageSize:
        description: |-
          PageSize 每页大小，选填，范围限制：[1,10000]
          用于限制每页返回的管理员数量，最小值为1，最大值为10000
        example: 10
        maximum: 10000
        minimum: 1
        type: integer
      recipeName:
        description: |-
          RecipeName string 配方名称，选填，长度限制：2-255字符
          用于按名称查询配方列表 前缀模糊匹配
        example: 配方名称
        maxLength: 255
        minLength: 2
        type: string
    type: object
  data.RecipeListResponse:
    properties:
      list:
        description: |-
          List []RecipeData 数据
          分页返回的配方列表
        items:
          $ref: '#/definitions/data.RecipeData'
        type: array
      total:
        description: |-
          Total int64 总条数
          返回符合条件的实验记录总数
        example: 100
        type: integer
    type: object
  dto.Empty:
    type: object
  dto.ErrorResponse:
    properties:
      code:
        description: 错误码
        type: integer
      message:
        description: 错误信息
        type: string
    type: object
info:
  contact: {}
paths:
  /ai/predict/forwardDirection:
    post:
      consumes:
      - application/json
      description: 提供配方信息并进行预测，返回实验结果的预测值
      parameters:
      - description: 配方预测请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ai.ForwardDirectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 预测成功，返回预测的实验结果
          schema:
            $ref: '#/definitions/ai.ForwardDirectionResponse'
        "400":
          description: 请求参数错误，可能是配方信息不完整或格式不正确
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是预测服务异常
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 根据配方信息预测实验结果
      tags:
      - 模型预测
...
  /data/experiment:
    delete:
      consumes:
      - application/json
      description: 根据实验ID删除实验
      parameters:
      - description: 请求参数，包含要删除的实验ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.ExperimentDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除实验
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，如实验ID不存在或格式无效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库删除失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 删除实验
      tags:
      - 实验管理
    get:
      consumes:
      - application/json
      description: 根据分页请求获取实验列表，支持按实验名称进行筛选
      parameters:
      - description: 请求参数，包含分页信息及筛选条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.ExperimentListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取实验列表，返回实验数据
          schema:
            $ref: '#/definitions/data.ExperimentListResponse'
        "400":
          description: 请求参数错误，如分页参数错误、筛选条件不符合要求
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库查询出错等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取实验列表
      tags:
      - 实验管理
    post:
      consumes:
      - application/json
      description: 根据提供的实验信息创建新的实验记录
      parameters:
      - description: 请求参数，包含新的实验信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.ExperimentAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加实验信息
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，如缺少必要字段或格式不正确
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库插入失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 添加新的实验信息
      tags:
      - 实验管理
    put:
      consumes:
      - application/json
      description: 根据实验ID修改实验的具体信息
      parameters:
      - description: 请求参数，包含实验ID及要修改的具体内容
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.ExperimentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修改实验信息
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，如实验ID不存在或修改内容无效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库更新失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 修改实验信息
      tags:
      - 实验管理
  /data/experiment/import:
    post:
      consumes:
      - multipart/form-data
      description: 接收上传的文件并处理，根据业务需求进行相关文件解析和导入
      parameters:
      - description: 上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 文件导入成功，返回空对象表示操作成功
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，可能是文件上传失败或格式不符合要求
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是文件解析或存储过程中出现异常
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 导入文件
      tags:
      - 实验管理
  /data/recipe:
    delete:
      consumes:
      - application/json
      description: 根据配方ID删除指定的配方记录
      parameters:
      - description: 请求参数，包含需要删除的配方ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除配方信息
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，可能是配方ID无效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库删除失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 删除配方
      tags:
      - 配方管理
    get:
      consumes:
      - application/json
      description: 根据分页请求获取配方列表，支持按配方名称进行筛选
      parameters:
      - description: 请求参数，包含分页信息及筛选条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取配方列表，返回配方数据
          schema:
            $ref: '#/definitions/data.RecipeListResponse'
        "400":
          description: 请求参数错误，如分页参数错误、筛选条件不符合要求
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库查询失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取配方列表
      tags:
      - 配方管理
    post:
      consumes:
      - application/json
      description: 根据提供的配方信息创建新的配方记录
      parameters:
      - description: 请求参数，包含新的配方信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加配方信息
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，如缺少必要字段或格式不正确
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库插入失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 添加新的配方
      tags:
      - 配方管理
    put:
      consumes:
      - application/json
      description: 根据提供的配方信息更新现有配方记录，更新时确保材料组和材料百分比的比例和为100%
      parameters:
      - description: 请求参数，包含更新的配方信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新配方信息
          schema:
            $ref: '#/definitions/dto.Empty'
        "400":
          description: 请求参数错误，如缺少必要字段或格式不正确
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库更新失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 编辑现有配方
      tags:
      - 配方管理
  /data/recipe/form/list:
    get:
      consumes:
      - application/json
      description: 根据分页请求获取配方表单列表，支持按配方名称进行筛选
      parameters:
      - description: 请求参数，包含分页信息及筛选条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeFormListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取配方表单列表，返回配方数据
          schema:
            $ref: '#/definitions/data.RecipeFormListResponse'
        "400":
          description: 请求参数错误，如分页参数错误、筛选条件不符合要求
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库查询失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取配方表单列表
      tags:
      - 配方管理
  /data/recipe/info:
    get:
      consumes:
      - application/json
      description: 根据传入的配方 ID 获取该配方的详细信息，包括材料组、材料信息以及基于此配方创建的实验信息
      parameters:
      - description: 请求参数，包含要查询的配方 ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/data.RecipeInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取单个配方详情，返回配方详细数据
          schema:
            $ref: '#/definitions/data.RecipeInfoResponse'
        "400":
          description: 请求参数错误，如配方 ID 格式错误等
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库查询失败等情况
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取单个配方详情
      tags:
      - 配方管理
...
swagger: "2.0"